moved applog to correct package
